# [NEON] Split XXH3 into 6 NEON lanes and 2 scalar lanes on aarch64