Instead of pure NEON, on AArch64 with size optimizations disabled, XXH3 will now use 6 lanes with NEON and 2 lanes with scalar by default.

This makes use of the otherwise inactive integer pipeline and results in massive speedup, especially with the previously underperforming GCC.

Note that this doesn't benefit ARMv7-a in the slightest.

This can be configured with the `XXH3_NEON_LANES` macro, which can be either 2, 4, 6, or 8.

**Google Pixel 4a** (2.21 GHz Snapdragon 730/Cortex-A76), `xxhsum -b`
|          |    Before   |     After    |  Diff. |
|:----|:-----|:-----|:------|
| GCC 11.1 | 7434.8 MB/s |  9814.9 MB/s | +32.0% |
| Clang 13 | 8788.4 MB/s | 10158.2 MB/s | +15.5% |