@easyaspi314 2021-12-01
And with that:
 - We finally break the 10 GB/s barrier on my phone
 - GCC is finally pulling its own weight
 - My phone now beats my i5-430m laptop (it gets 9.8 GB/s on Clang 13)

Unfortunately, this doesn't seem to affect x86_64 in any positive manner.