@Cyan4973 2021-12-01
Somehow, this feels like a distant cousin of multi-cores processing, though at a more granular instruction port level.
It's nice and interesting, I like it.