@easyaspi314 2021-12-01
Should I move the scalar variant to the top?