@Cyan4973 2021-12-01
> Should I move the scalar variant to the top?

So that they stand closer to their component `XXH3_scalar*Round()` ?
Yes, maybe but this can wait another PR.