@easyaspi314 2021-12-01
Well I am already moving the implementation to the top because I need to use it in both, and it isn't used as the last `#else` case.