@Cyan4973 2021-12-01
You could also declare them as the top, and keep their implementation in the scalar category.

Whatever seems simpler / more readable.