@easyaspi314 2021-12-01
> You could also declare them as the top, and keep their implementation in the scalar category.
> 
> Whatever seems simpler / more readable.

I did the forward deck for now.

I also added some explanations of ARM's uop dispatcher for why it works.

Or, if you want the TL;DR:
![arm all pipelines.jpg](https://user-images.githubusercontent.com/6258309/144202898-e4aacb48-ca1b-4197-ba12-4123d2c51a0c.jpg)