@easyaspi314 2021-12-01
I should test this on the in-order base model Cortex-A53. My mom has a Verizon Ellipsis 10 with one.

But yeah I'll make a new PR with some new documentation.