@easyaspi314 2021-12-01
Cortex-A53 test results

"Mom, can I have an AArch64 CPU?"
"We already have an AArch64 CPU at home."
The AArch64 CPU at home:
```
$ ./xxhsum-full-neon -b
xxhsum-full-neon 0.8.1 by Yann Collet 
compiled as 64-bit aarch64 + NEON little endian with Clang 13.0.1  
Sample of 100 KB...        
 1#XXH32                         :     102400 ->     7130 it/s (  696.3 MB/s)   
 3#XXH64                         :     102400 ->    17398 it/s ( 1699.0 MB/s)   
 5#XXH3_64b                      :     102400 ->    19757 it/s ( 1929.4 MB/s)   
11#XXH128                        :     102400 ->    19737 it/s ( 1927.4 MB/s)   
$ ./xxhsum-6-neon -b
xxhsum-6-neon 0.8.1 by Yann Collet 
compiled as 64-bit aarch64 + NEON little endian with Clang 13.0.1  
Sample of 100 KB...        
 1#XXH32                         :     102400 ->     7128 it/s (  696.1 MB/s)   
 3#XXH64                         :     102400 ->    17400 it/s ( 1699.2 MB/s)   
 5#XXH3_64b                      :     102400 ->    19918 it/s ( 1945.1 MB/s)   
11#XXH128                        :     102400 ->    19907 it/s ( 1944.1 MB/s)   
```