@Cyan4973 2021-12-01
meaning, `XXH32` is rather slower than expected ?