@easyaspi314 2021-12-02
> meaning, `XXH32` is rather slower than expected ?

Actually, I forgot that XXH32 was slower on the older CPUs lol. Although I do see a slight optimization I can make in load/store.

I also tested this PR on my old Pixel 2 XL (Snap 835/Cortex-A73). This has the older 3 micro-op dispatch buffer, and only gets a ~4% speedup from this (5.1->5.3 GB/s).

So for the three main dispatch styles, we don't ever lose performance, as I predicted.
 - In-order dual issue (A53): no change
 - OoO 3 micro-op (A73): slight improvement
 - OoO 8 micro-op (A76): big improvement